+++ src/kate-client/store.js
@@ -3,12 +3,12 @@ import logger from 'redux-logger';
 
 import rootReducer from './reducers';
 
-let middleware;
-if (process.env.NODE_ENV === 'development') {
-  middleware = applyMiddleware(logger);
-}
+export default function configureStore(initialState, useLogger) {
+  let middleware;
+  if (useLogger) {
+    middleware = applyMiddleware(logger);
+  }
 
-export default function configureStore(initialState) {
   const store = createStore(rootReducer, initialState, middleware);
   return store;
 }